+++ lib/docker_commander.dart
@@ -1,6 +1,7 @@
 library docker_commander;
 
 export 'src/docker_commander_base.dart';
+export 'src/docker_commander_console.dart';
 export 'src/docker_commander_containers.dart';
 export 'src/docker_commander_host.dart';
 export 'src/docker_commander_nginx.dart';